🔨🤖 remove react-refresh lint suppressions, split up mixed exports
Drop the seven eslint-disable comments for react/only-export-components
and make the files comply instead:

- un-export ProminentLinkStyles (only used internally)
- move showChartType into ChartRow.tsx, its only consumer
- delete the SearchField alias class; use TextField directly
- extract EditorDatabase (+ Dataset/Namespace types) into EditorDatabase.ts
- extract IconToggleComponent from GrapherConfigGridEditorTypesAndUtils,
  which becomes a component-free .ts file
- move runFeedbackPage into FeedbackPageMain.tsx
- move hydrateDynamicCollectionPage and embedDynamicCollectionGrapher
  into DynamicCollectionPageMain.tsx
