+++ adminSiteClient/ChartEditorView.tsx
@@ -1,4 +1,3 @@
-/* eslint-disable react/only-export-components */
 import * as _ from "lodash-es"
 import * as React from "react"
 import { observer } from "mobx-react"
@@ -62,48 +61,7 @@ import {
     ErrorMessagesForDimensions,
     FieldWithDetailReferences,
 } from "./ChartEditorTypes.js"
-
-interface Variable {
-    id: number
-    name: string
-}
-
-export interface Dataset {
-    id: number
-    name: string
-    namespace: string
-    version: string | undefined
-    variables: Variable[]
-    isPrivate: boolean
-    nonRedistributable: boolean
-}
-
-export interface Namespace {
-    name: string
-    description?: string
-    isArchived: boolean
-}
-
-// This contains the dataset/variable metadata for the entire database
-// Used for variable selector interface
-export interface NamespaceData {
-    datasets: Dataset[]
-}
-
-export class EditorDatabase {
-    namespaces: Namespace[]
-    variableUsageCounts: Map<number, number> = new Map()
-    dataByNamespace: Map<string, NamespaceData> = new Map()
-
-    constructor(json: any) {
-        makeObservable(this, {
-            namespaces: observable.ref,
-            variableUsageCounts: observable.ref,
-            dataByNamespace: observable,
-        })
-        this.namespaces = json.namespaces
-    }
-}
+import { Dataset, EditorDatabase } from "./EditorDatabase.js"
 
 export type DetailReferences = Record<FieldWithDetailReferences, string[]>
