+++ adminSiteClient/EditorDatabase.ts
@@ -1,0 +1,43 @@
+import { observable, makeObservable } from "mobx"
+
+interface Variable {
+    id: number
+    name: string
+}
+
+export interface Dataset {
+    id: number
+    name: string
+    namespace: string
+    version: string | undefined
+    variables: Variable[]
+    isPrivate: boolean
+    nonRedistributable: boolean
+}
+
+export interface Namespace {
+    name: string
+    description?: string
+    isArchived: boolean
+}
+
+// This contains the dataset/variable metadata for the entire database
+// Used for variable selector interface
+export interface NamespaceData {
+    datasets: Dataset[]
+}
+
+export class EditorDatabase {
+    namespaces: Namespace[]
+    variableUsageCounts: Map<number, number> = new Map()
+    dataByNamespace: Map<string, NamespaceData> = new Map()
+
+    constructor(json: any) {
+        makeObservable(this, {
+            namespaces: observable.ref,
+            variableUsageCounts: observable.ref,
+            dataByNamespace: observable,
+        })
+        this.namespaces = json.namespaces
+    }
+}
