+++ adminSiteClient/IconToggleComponent.tsx
@@ -1,0 +1,18 @@
+import type { IconDefinition } from "@fortawesome/fontawesome-common-types"
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome"
+
+export interface IconToggleProps {
+    isOn: boolean
+    onIcon: IconDefinition
+    offIcon: IconDefinition
+    onClick: (newState: boolean) => void
+}
+
+export const IconToggleComponent = (props: IconToggleProps) => (
+    <button
+        className="btn btn-light btn-sm"
+        onClick={() => props.onClick(!props.isOn)}
+    >
+        <FontAwesomeIcon icon={props.isOn ? props.onIcon : props.offIcon} />
+    </button>
+)
