+++ site/Feedback.tsx
@@ -1,4 +1,3 @@
-/* eslint-disable react/only-export-components */
 import * as React from "react"
 import { observer } from "mobx-react"
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome"
@@ -11,7 +10,6 @@ import { observable, action, toJS, computed, makeObservable } from "mobx"
 import classnames from "clsx"
 import { BAKED_BASE_URL } from "../settings/clientSettings.js"
 import { stringifyUnknownError } from "@ourworldindata/utils"
-import { createRoot } from "react-dom/client"
 
 const sendFeedback = async (feedback: Feedback) => {
     const json = {
@@ -468,16 +466,4 @@ export class FeedbackPrompt extends React.Component {
             </div>
         )
     }
-}
-
-export function runFeedbackPage() {
-    const elem = document.querySelector(".FeedbackPage main")
-    if (!elem) throw new Error("FeedbackPage main element not found in DOM")
-
-    const root = createRoot(elem)
-    root.render(
-        <div className="box">
-            <FeedbackForm />
-        </div>
-    )
 }
