+++ site/owid.entry.ts
@@ -15,7 +15,7 @@ import "@fortawesome/fontawesome-svg-core/styles.css"
 import React from "react"
 import { createRoot } from "react-dom/client"
 import { runNotFoundPage } from "./NotFoundPageMain.js"
-import { runFeedbackPage } from "./Feedback.js"
+import { runFeedbackPage } from "./FeedbackPageMain.js"
 import { runDonateForm } from "./runDonateForm.js"
 import { Explorer } from "@ourworldindata/explorer"
 import { ENV } from "../settings/clientSettings.js"
